Refactoring (#810)
* - Add another optimization variation + fixed typo

* - More Optimizations

* - More Optimizations

* - Progress

* - New IR instructions to separate object references (part of the GC plan)

* - Progress

* - Progress

* - Progress

* - Crushing invisible bugs (due to more validation checks)

* - Crushing invisible bugs (due to more validation checks)

* - Crushing invisible bugs (due to more validation checks)

* - Crushing invisible bugs (due to more validation checks)

* - A few more optimizations.

* - Re-arrange source code

* - Refactoring Optimization stages

* - Refactoring Optimization stages

* - Refactoring

* - Refactoring

* - Merged Floating Point stage

* - Cleanup

* - Cleanup

* - Cleanup

* - Cleanup

* - Merged 32 and 64 transformations stages

* - Merged 32 and 64 transformations stages

* - Merged 32 and 64 transformations stages

+++ Source/Mosa.Compiler.Framework/CompilerStages/PreLinkHashFileStage.cs
@@ -76,7 +76,7 @@ namespace Mosa.Compiler.Framework.CompilerStages
 				}
 			}
 
-			info.Sort((HashInfo x, HashInfo y) => (int)string.Compare(y.Name, x.Name));
+			info.Sort((HashInfo x, HashInfo y) => string.Compare(y.Name, x.Name));
 
 			return info;
 		}
