+++ Source/Mosa.Compiler.Framework/Linker/Elf/RelocationAddendEntry.cs
@@ -76,9 +76,9 @@ namespace Mosa.Compiler.Framework.Linker.Elf
 		/// <param name="writer">The writer.</param>
 		protected void Write64(BinaryWriter writer)
 		{
-			writer.Write((ulong)Offset);
+			writer.Write(Offset);
 			writer.Write((ulong)Info64);
-			writer.Write((ulong)Addend);
+			writer.Write(Addend);
 		}
 	}
 }
